requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
banks==2.4.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.1
dirtyjson==1.0.8
distro==1.9.0
docling==2.84.0
docling-core==2.71.0
docling-ibm-models==3.13.0
docling-parse==5.7.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
email-validator==2.2.0
exceptiongroup==1.3.1
Faker==40.12.0
fastapi==0.135.3
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.70.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langextract==1.2.0
latex2mathml==3.79.0
llama-index==0.14.19
llama-index-cli==0.5.6
llama-index-core==0.14.19
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.5
llama-index-workflows==2.17.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
ml_collections==1.1.0
mmh3==5.2.1
more-itertools==10.8.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
ocrmac==1.0.1
omegaconf==2.3.0
openai==2.30.0
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
bcrypt==5.0.0
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.4
pluggy==1.6.0
polyfactory==3.3.0
postgrest==2.28.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pylatexenc==2.10
PyMySQL==1.1.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Vision==12.1
pyodbc==5.3.0
pyparsing==3.3.2
pypdfium2==5.6.0
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
PyYAML==6.0.3
rapidocr==3.7.0
realtime==2.28.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
safetensors==0.7.0
scipy==1.17.1
semchunk==3.2.5
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.4
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
stripe==15.0.1
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.4.0
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==15.0.1
wrapt==2.1.2
xlsxwriter==3.2.9
yarl==1.23.0
zstandard==0.25.0

# Fase 2 — Connectors
msal==1.31.1
slack-sdk==3.33.5
simple-salesforce==1.12.6
paramiko==3.5.1
